Create readability.c
This is from cs50 class probelm set 2, readability. If you inputs some texts, it will analyze your text and tells whether it's hard to read or not by printing appropriate grade (age) to read. Here is some random text! 
The term atomic physics can be associated with nuclear power and nuclear weapons, due to the synonymous use of atomic and nuclear in standard English. Physicists distinguish between atomic physics—which deals with the atom as a system consisting of a nucleus and electrons—and nuclear physics, which studies nuclear reactions and special properties of atomic nuclei. (Wikipedia - Atomic physics) 
It should print "Grade 16"